requirements.txt
-e git+https://github.com/bufferapp/buffer-python.git#egg=buffer-python
emoji==0.3.4
ipython==3.1.0
python-daemon==2.0.5
slackclient==0.15
smartypants==1.8.6
twitter-text-py==2.0.2
